+++ src/nal/slice/mod.rs
@@ -1,3 +1,7 @@
+mod cavlc;
+pub mod data;
+pub mod macroblock;
+
 use crate::nal::pps;
 use crate::nal::pps::{PicParamSetId, PicParameterSet};
 use crate::nal::sps;